Fix Unlink during Remove and Update File

+++ Entity/Banner.php
@@ -319,8 +319,10 @@ class Banner
 
         // check if we have an old image
         if (isset($this->temp)) {
-            // delete the old image
-            unlink($this->getUploadRootDir().$this->temp);
+            if (file_exists($this->getUploadRootDir().$this->temp) && is_file($this->getUploadRootDir().$this->temp)) {
+                // delete the old image
+                unlink($this->getUploadRootDir().$this->temp);
+            }
             // clear the temp image path
             $this->temp = null;
         }
@@ -333,7 +335,7 @@ class Banner
     public function removeUpload()
     {
         $file = $this->getAbsolutePath();
-        if ($file) {
+        if (file_exists($file) && is_file($file)) {
             unlink($file);
         }
     }
