+++ Entity/News.php
@@ -374,8 +374,10 @@ class News
 
         // check if we have an old image
         if (isset($this->temp)) {
-            // delete the old image
-            unlink($this->getUploadRootDir().$this->temp);
+            if (file_exists($this->getUploadRootDir().$this->temp) && is_file($this->getUploadRootDir().$this->temp)) {
+                // delete the old image
+                unlink($this->getUploadRootDir().$this->temp);
+            }
             // clear the temp image path
             $this->temp = null;
         }
@@ -388,7 +390,7 @@ class News
     public function removeUpload()
     {
         $file = $this->getAbsolutePath();
-        if ($file) {
+        if (file_exists($file) && is_file($file)) {
             unlink($file);
         }
     }
